+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Make option `--modify` obsolete ([#87](https://github.com/josefpihrt/orang/pull/87)).
 
+### Fixed
+
+- [CLI] Fix reading of redirected input ([PR](https://github.com/dotnet/roslynator/pull/91))
+
 ## [0.6.0] - 2023-11-07
 
 ## [0.6.0-beta] - 2023-10-19